requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.4
decorator==5.1.1
Django==4.2.5
django-cors-headers==4.4.0
django-filter==24.3
django-rest-knox==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
frozenlist==1.4.1
gunicorn==20.0.4
idna==3.6
inflection==0.5.1
kombu==5.3.7
multidict==6.0.5
packaging==24.1
Pillow==10.0.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.5
py==1.11.0
pycparser==2.21
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
retry==0.9.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
